tilda-whatsapp-bot/requirements.txt
python-telegram-bot==20.6
Flask==2.3.2

python-dotenv==1.0.1
